Prevent weirdness when the user accidentally selects a directory
The `ln` command will treat the symbolic link as if you wish to use the
form where you make symbolic links in a directory with the same name as
the target.  This is confusing when it happens until you realize what
is happening.  The `-T` switch prevents this from happening.